refactor: 테스트 파일 리팩토링
- 반복적으로 사용되는 mockRandoms, mockQuestion,getLogSpy를 testUtls/index.js 로 이동
