+++ __tests__/CarTest.js
@@ -1,24 +1,15 @@
-import { MissionUtils } from '@woowacourse/mission-utils';
 import Car from '../src/Car.js';
+import { mockRandoms } from '../testUtils/index.js';
 
-const mockRandoms = (numbers) => {
-  MissionUtils.Random.pickNumberInRange = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, MissionUtils.Random.pickNumberInRange);
-};
-
-
-describe('자동차 게임:자동차 테스트',()=>{
-
-  test("자동차 이름",()=>{
-    const name ="pobi";
+describe('자동차 게임:자동차 테스트', () => {
+  test('자동차 이름', () => {
+    const name = 'pobi';
     const car = new Car(name);
     expect(car.name).toBe(name);
   });
 
-  test("자동차 전진",()=>{
-    const car = new Car("pobi");
+  test('자동차 전진', () => {
+    const car = new Car('pobi');
     mockRandoms([3]);
     car.handleMovement();
     expect(car.movement.length).toBe(0);
