+++ __tests__/MessageTest.js
@@ -1,30 +1,8 @@
 import App from '../src/App.js';
 
-import { MissionUtils } from '@woowacourse/mission-utils';
+import { mockQuestions, mockRandoms, getLogSpy } from '../testUtils';
 
 import { MESSAGE } from '../src/constant/';
-
-const mockQuestions = (inputs) => {
-  MissionUtils.Console.readLineAsync = jest.fn();
-
-  MissionUtils.Console.readLineAsync.mockImplementation(() => {
-    const input = inputs.shift();
-    return Promise.resolve(input);
-  });
-};
-
-const mockRandoms = (numbers) => {
-  MissionUtils.Random.pickNumberInRange = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, MissionUtils.Random.pickNumberInRange);
-};
-
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
-  logSpy.mockClear();
-  return logSpy;
-};
 
 describe('자동차 경주 게임: 출력 메세지 테스트', () => {
   test('입력에 대한 메세지 출력 및 유효한 입력값 출력', async () => {
