+++ __tests__/ValidateTest.js
@@ -1,23 +1,12 @@
 import App from '../src/App.js';
 import { MESSAGE } from '../src/constant';
-import { MissionUtils } from '@woowacourse/mission-utils';
-
-const mockQuestions = (inputs) => {
-  MissionUtils.Console.readLineAsync = jest.fn();
-
-  MissionUtils.Console.readLineAsync.mockImplementation(() => {
-    const input = inputs.shift();
-    return Promise.resolve(input);
-  });
-};
-
-const getLogSpy = () => {
-  const logSpy = jest.spyOn(MissionUtils.Console, 'print');
-  logSpy.mockClear();
-  return logSpy;
-};
+import { mockQuestions, getLogSpy } from '../testUtils/index.js';
 
 describe('자동차 경주 게임: 입력값 유효성 검사', () => {
+  let app;
+  beforeEach(() => {
+    app = new App();
+  });
   test.each([
     [['po비,javii', '0']],
     [
@@ -28,7 +17,6 @@ describe('자동차 경주 게임: 입력값 유효성 검사', () => {
     mockQuestions(inputs);
 
     const logSpy = getLogSpy(jest);
-    const app = new App();
     await app.play();
 
     expect(logSpy).toHaveBeenCalledWith(
@@ -41,8 +29,6 @@ describe('자동차 경주 게임: 입력값 유효성 검사', () => {
     async (inputs) => {
       mockQuestions(inputs);
 
-      const app = new App();
-
       await expect(app.play()).rejects.toThrow(MESSAGE.nameError);
     }
   );
@@ -53,7 +39,7 @@ describe('자동차 경주 게임: 입력값 유효성 검사', () => {
     [['pobi,woni', '하나']],
   ])('이동 횟수 대한 예외 처리: 0이상의 숫자만 가능', async (inputs) => {
     mockQuestions(inputs);
-    const app = new App();
+
     await expect(app.play()).rejects.toThrow(MESSAGE.roundError);
   });
 });
